Fix operator parsing - clean
