+++ src/phase/generator/function.salam
@@ -7,6 +7,8 @@
     نمایش ۱۰ - ۴ * ۲
     نمایش ۲ ** ۳ + ۴
 
+    نمایش -۴
+
 //    نمایش ۴ + 6
 //    برگشت ۴
 تمام
