+++ src/phase/parser/parse/ast/expression/binary.ts
@@ -1,12 +1,13 @@
 import { AstExpression } from './expression';
 
-export class AstBinaryExpression extends AstExpression {
+export class AstExpressionBinary extends AstExpression {
     left: AstExpression;
     operator: string;
     right: AstExpression;
 
     constructor(left: AstExpression, operator: string, right: AstExpression) {
-        super(left.getString() + " " + operator + " " + right.getString());
+        super("ExpressionBinary");
+        // left.getString() + " " + operator + " " + right.getString()
         this.left = left;
         this.operator = operator;
         this.right = right;
@@ -14,7 +15,7 @@ export class AstBinaryExpression extends AstExpression {
 
     stringify(wantsJson: boolean = true): string | object {
         return {
-            type: "BinaryExpression",
+            type: "ExpressionBinary",
             operator: this.operator,
             left: this.left.stringify(wantsJson),
             right: this.right.stringify(wantsJson),
