+++ src/phase/parser/parse/ast/expression/unary.ts
@@ -1,18 +1,19 @@
 import { AstExpression } from './expression';
 
-export class AstUnaryExpression extends AstExpression {
+export class AstExpressionUnary extends AstExpression {
     operator: string;
     operand: AstExpression;
 
     constructor(operator: string, operand: AstExpression) {
-        super(operator + " " + operand.getString());
+        super("ExpressionUnary");
+        // operator + " " + operand.getString()
         this.operator = operator;
         this.operand = operand;
     }
 
     stringify(wantsJson: boolean = true): string | object {
         return {
-            type: "UnaryExpression",
+            type: "ExpressionUnary",
             operator: this.operator,
             operand: this.operand.stringify(wantsJson),
         };
