+++ src/phase/parser/parse/statement/statement.ts
@@ -1,12 +1,11 @@
 import { Parser } from './../parser';
-import { AstBlock } from './../ast/block';
-import { Token } from './../../../lexer/tokenizer/token';
 import { parseStatementPrint } from './print';
 import { parseStatementReturn } from './return';
+import { Token } from './../../../lexer/tokenizer/token';
+import { AstStatement } from './../ast/statement/statement';
 
-export function parseStatement(parser: Parser): AstBlock | undefined {
+export function parseStatement(parser: Parser): AstStatement | undefined {
     const current_token: Token | undefined = parser.currentToken;
-    console.log("parseStatement: ", current_token.type);
     switch (current_token.type) {
         case "RET": return parseStatementReturn(parser);
         case "PRINT": return parseStatementPrint(parser);
